Many bugfixes; add profile select feature to prefsCleaner.sh.
This commit brings many bugfixes after more thorough testing across
all targeted platforms (except macOS) and shells. Some of the compound
commands in the POSIX shell language, e.g. the while and for loops,
have been restructured to allow for more error conditions to be caught.
Variable names are reused more often to avoid needlessly littering the
shell environment, as they are not scoped locally to the functions,
and to reduce the chance of a name collision with variables that are
marked readonly for whatever reason.